+++ tools/tesseratos/src/tesseratos/world_inspector/plugin.cpp
@@ -4,6 +4,7 @@
 #include <cubos/core/ecs/query/opt.hpp>
 
 #include <cubos/engine/imgui/plugin.hpp>
+#include <cubos/engine/transform/child_of.hpp>
 
 #include <tesseratos/entity_selector/plugin.hpp>
 #include <tesseratos/toolbox/plugin.hpp>
@@ -14,8 +15,55 @@ using cubos::core::ecs::Name;
 using cubos::core::ecs::Opt;
 using cubos::core::ecs::Query;
 using cubos::core::ecs::World;
+using cubos::core::ecs::WriteResource;
+using cubos::engine::ChildOf;
 
 using cubos::engine::Cubos;
+
+static void show(Entity entity, Query<Entity, Opt<const Name&>, const ChildOf&, Entity> childOf, Opt<const Name&> name,
+                 tesseratos::EntitySelector& selector)
+{
+    auto displayName = name.contains() ? name->value : std::to_string(entity.index);
+    auto children = childOf.pin(1, entity);
+    auto textColor = (ImVec4)ImColor(255, 255, 255);
+
+    selector.selection == entity ? textColor = (ImVec4)ImColor(149, 252, 75)
+                                 : textColor = (ImVec4)ImColor(255, 255, 255);
+
+    // check if entity has children
+    if (children.empty())
+    {
+        ImGui::PushStyleColor(ImGuiCol_Text, textColor);
+        ImGui::Selectable(("   " + displayName).c_str(), selector.selection == entity);
+        ImGui::PopStyleColor();
+
+        if (ImGui::IsItemClicked(0))
+        {
+            selector.selection = entity;
+        }
+
+        return;
+    }
+
+    ImGui::PushStyleColor(ImGuiCol_Text, textColor);
+    if (ImGui::TreeNode(displayName.c_str()))
+    {
+        ImGui::PopStyleColor();
+
+        if (ImGui::IsItemClicked(0))
+        {
+            selector.selection = entity;
+        }
+        for (auto [child, nameChild, _, parent] : children)
+        {
+            show(child, childOf, nameChild, selector);
+        }
+        ImGui::TreePop();
+
+        return;
+    }
+    ImGui::PopStyleColor();
+}
 
 void tesseratos::worldInspectorPlugin(Cubos& cubos)
 {
@@ -24,7 +72,8 @@ void tesseratos::worldInspectorPlugin(Cubos& cubos)
     cubos.addPlugin(entitySelectorPlugin);
     cubos.system("show World Inspector UI")
         .tagged("cubos.imgui")
-        .call([](World& world, Query<Entity, Opt<const Name&>> query) {
+        .call([](World& world, Query<Entity, Opt<const Name&>> all,
+                 Query<Entity, Opt<const Name&>, const ChildOf&, Entity> query) {
             if (!(world.write<Toolbox>().get().isOpen("World Inspector")))
             {
                 return;
@@ -35,28 +84,14 @@ void tesseratos::worldInspectorPlugin(Cubos& cubos)
             ImGui::Begin("World Inspector");
             if (!ImGui::IsWindowCollapsed())
             {
-                int n = 0;
-                for (auto [entity, name] : query)
+                for (auto [entity, name] : all)
                 {
-                    ImGui::PushID(n);
-
-                    if (name.contains())
+                    if (!query.pin(0, entity).empty())
                     {
-                        ImGui::BulletText("%s", name->value.c_str());
-                    }
-                    else
-                    {
-                        ImGui::BulletText("%s", std::to_string(entity.index).c_str());
+                        continue; // Skip all entities with parents
                     }
 
-                    ImGui::SameLine();
-                    if (ImGui::Button("Select"))
-                    {
-                        selector.get().selection = entity;
-                    }
-
-                    n++;
-                    ImGui::PopID();
+                    show(entity, query, name, selector.get());
                 }
             }
             ImGui::End();